Create kmean_MINST.py
In this code, we first load the MNIST dataset using the fetch_openml function from scikit-learn. We then split the dataset into training and testing sets using the train_test_split function.

Next, we extract the centroid feature using the KMeans algorithm from scikit-learn. We fit a KMeans model with 10 clusters to the training data, and then transform both the training and testing data to obtain their respective centroid representations.

Finally, we train a KNN classifier with 5 neighbors on the centroid feature, and evaluate its accuracy on the testing data.